requirements.txt
alembic==1.13.1
aniso8601==9.0.1
apispec==6.4.0
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==6.5.0
coveralls==3.3.1
docopt==0.6.2
Flask==3.0.2
flask-apispec==0.11.4
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
packaging==23.2
pluggy==1.4.0
psycopg2==2.9.9
pytest==8.0.1
python-coveralls==2.9.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==2.2.1
webargs==8.4.0
Werkzeug==3.0.1
sqlalchemy_utils==0.41.1
